Begin preparing the mod for Update 8
This commit encompasses the following changes:

- Bump mod version to `1.1.0`
- Increase SML dep to `^3.7.0` and enable dedi server targets
- Stop using beacon assets as they are slated for removal
- Add stub events for primary and secondary fire input actions
- Add 3rd person equipment attachment, based on vanilla cup
- Re-do materials to use Panini projection (1st person view)
- Add waves and translucency to make cup contents liquid-like
- Use a single texture for all in-game icons (128 x 128)
- Reduce mesh triangle count to 25% (could be lowered further)
- Make conveyor mesh significantly larger to ease picking it up
- Add in-game mod menu icon (the same icon as the mod page)
- Move recipe to the Equipment Workshop (the cup *is* equipment)

IMPORTANT NOTE: To make the Panini projection work, one needs to have
https://github.com/satisfactorymodding/SatisfactoryModLoader/pull/262
applied to the Starter Project. In addition, to use `MPC_Panini`, one
needs to replace the `MPC_Panini.uasset` file in Starter Project with
the `.uasset` and `.uexp` files from FModel and enable cooked content
in the editor, in `Config/DefaultEngine.ini`:

    [/Script/UnrealEd.CookerSettings]
    cook.AllowCookedDataInEditorBuilds=True
    s.AllowUnversionedContentInEditor=1

Using the cooked MPC asset is most certainly not ideal, but it is the
only known (to date) workaround to reference MPC's from within a mod.

This commit has some known "bugs" (or features):

- Red solo cup does not have any animations or sounds
- Cup slightly clips into hand (reuses Xeno-Basher "gripping")
- The contents of the cup do not obey the laws of physics

Signed-off-by: Angel Pons <th3fanbus@gmail.com>